xdc: cast CVErrorCodes to int for fmt 11 compatibility
HidraQTPDProducer formats the CAEN `CVErrorCodes` enum directly with fmt's `{}`
in two HIDRA_ERROR calls. fmt v11 removed the implicit enum-to-int conversion,
so this no longer compiles (`type_is_unformattable_for<CVErrorCodes>`). Cast the
enum to int at the two call sites; works on all fmt versions.

+++ user/hidra/xdc/src/HidraQTPDProducer.cc
@@ -841,13 +841,13 @@ void WriteEventSyncTrigger16(uint64_t triggerNumber) {
 
     if (byteCount <= 0) {
       std::this_thread::sleep_for(std::chrono::milliseconds(1));
-      HIDRA_ERROR("BCNT = 0, controller replied with 0 bytes. Ret code of BLT was {}", ret);
+      HIDRA_ERROR("BCNT = 0, controller replied with 0 bytes. Ret code of BLT was {}", static_cast<int>(ret));
       return false;
     }
 
     const int wordCount = byteCount / 4;
     if (wordCount <= 0) {
-      HIDRA_ERROR("BCNT = 0, controller replied with less than 4 bytes. Ret code of BLT was {}", ret);
+      HIDRA_ERROR("BCNT = 0, controller replied with less than 4 bytes. Ret code of BLT was {}", static_cast<int>(ret));
       return false;
     }
